TF-BOSS/requirements.txt
absl-py
easydict
cython
numpy
#tensorflow-gpu==1.14.0
tqdm
